Hack around the "cannot move out of borrowed context" with Clone
Seems weird that I have to clone the User out of the Vec<User> to be
able to make an immutable reference to one of its members. There's
almost definitely a better way to do this depending on the context.
Splitting out the main function into smaller functions such that the
same set of references don't share the same scope would likely fix this
too.